Moved src to group package
Closes #9
