+++ src/main/kotlin/com/github/agcom/quadtree/quadtree/geometry/Area.kt
@@ -1,0 +1,9 @@
+package com.github.agcom.quadtree.quadtree.geometry
+
+interface Area {
+
+    operator fun contains(p: Point): Boolean
+
+    fun overlaps(rect: Rectangle): Boolean
+
+}
