+++ src/main/kotlin/com/github/agcom/quadtree/quadtree/geometry/Point.kt
@@ -1,0 +1,3 @@
+package com.github.agcom.quadtree.quadtree.geometry
+
+data class Point(val x: Int, val y: Int)
