--- src/main/kotlin/quadtree/geometry/Area.kt
@@ -1,9 +1,0 @@
-package quadtree.geometry
-
-interface Area {
-
-    operator fun contains(p: Point): Boolean
-
-    fun overlaps(rect: Rectangle): Boolean
-
-}
