--- src/main/kotlin/quadtree/geometry/Point.kt
@@ -1,3 +1,0 @@
-package quadtree.geometry
-
-data class Point(val x: Int, val y: Int)
