+++ src/main/resources/com/github/agcom/quadtree/images/launcher.png
