+++ src/main/resources/com/github/agcom/quadtree/layouts/home.fxml
@@ -1,0 +1,114 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import java.lang.*?>
+<?import java.util.*?>
+<?import javafx.scene.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+
+<?import com.jfoenix.controls.JFXTextField?>
+<?import com.github.agcom.quadtree.nodes.IntField?>
+<?import java.net.URL?>
+<?import com.jfoenix.controls.JFXTextArea?>
+<?import com.jfoenix.controls.JFXRadioButton?>
+<?import com.github.agcom.quadtree.nodes.NodeSwitcher?>
+<?import com.jfoenix.controls.JFXCheckBox?>
+<?import com.github.agcom.quadtree.nodes.LabeledSeparator?>
+<?import com.github.agcom.quadtree.nodes.FakeFocus?>
+<SplitPane xmlns="http://javafx.com/javafx"
+           xmlns:fx="http://javafx.com/fxml"
+           fx:controller="com.github.agcom.quadtree.controllers.HomeController"
+           prefHeight="800" prefWidth="1200"
+           dividerPositions="0.3">
+
+    <stylesheets>
+
+        <URL value="@/com/github/agcom/quadtree/styles/global.css"/>
+        <URL value="@/com/github/agcom/quadtree/styles/home.css"/>
+
+    </stylesheets>
+
+    <FakeFocus/>
+
+    <VBox minWidth="300" fx:id="controlPane" spacing="32" SplitPane.resizableWithParent="true">
+
+        <LabeledSeparator text="Quad Tree Specs"/>
+
+        <VBox spacing="24">
+
+            <HBox spacing="16">
+
+                <IntField fx:id="width" labelFloat="true" promptText="Width" text="600" HBox.hgrow="ALWAYS"/>
+                <IntField fx:id="height" labelFloat="true" promptText="Height" text="600" HBox.hgrow="ALWAYS"/>
+
+            </HBox>
+
+            <IntField fx:id="capacity" labelFloat="true" promptText="Capacity" text="4"/>
+
+            <IntField fx:id="randomize" promptText="Random points" labelFloat="true" text="100" onAction="#randomize"/>
+
+        </VBox>
+
+        <LabeledSeparator text="Query options"/>
+
+        <VBox spacing="24">
+
+            <fx:define>
+
+                <ToggleGroup fx:id="queryShapeGroup"/>
+
+            </fx:define>
+
+            <HBox spacing="16">
+
+                <JFXRadioButton toggleGroup="$queryShapeGroup" text="Rectangle" userData="rect" fx:id="rectQuery"
+                                selected="true"/>
+                <JFXRadioButton toggleGroup="$queryShapeGroup" text="Circle" userData="circle" fx:id="circleQuery"/>
+
+            </HBox>
+
+            <NodeSwitcher fx:id="queryOptionsSwitcher" startNodeIndex="0">
+
+                <nodes>
+
+                    <HBox spacing="16">
+
+                        <IntField fx:id="queryWidth" labelFloat="true" promptText="Width" text="200"
+                                  HBox.hgrow="ALWAYS"/>
+                        <IntField fx:id="queryHeight" labelFloat="true" promptText="Height" text="200"
+                                  HBox.hgrow="ALWAYS"/>
+
+                    </HBox>
+
+                    <IntField fx:id="radius" labelFloat="true" promptText="Radius" text="100"/>
+
+                </nodes>
+
+            </NodeSwitcher>
+
+            <JFXCheckBox fx:id="eraser" text="Eraser"/>
+
+        </VBox>
+
+        <LabeledSeparator text="Delays"/>
+
+        <HBox spacing="16">
+
+            <IntField fx:id="drawDelay" promptText="Draw delay" labelFloat="true" text="0" HBox.hgrow="ALWAYS"/>
+
+            <IntField fx:id="queryDelay" promptText="Query delay" labelFloat="true" text="10" HBox.hgrow="ALWAYS"/>
+
+            <IntField fx:id="rebuildDelay" promptText="Rebuild delay" labelFloat="true" text="1000"
+                      HBox.hgrow="ALWAYS"/>
+
+        </HBox>
+
+        <LabeledSeparator text="Query statistics"/>
+
+        <TextArea fx:id="logs" editable="false" VBox.vgrow="ALWAYS"/>
+
+    </VBox>
+
+    <fx:include source="quadtree.fxml" fx:id="quadTree" SplitPane.resizableWithParent="true"/>
+
+</SplitPane>
