+++ src/main/resources/com/github/agcom/quadtree/styles/global.css
