--- src/main/resources/layouts/quadtree.fxml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import java.lang.*?>
-<?import java.util.*?>
-<?import javafx.scene.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-
-<?import javafx.scene.canvas.Canvas?>
-<?import javafx.scene.shape.Shape?>
-<BorderPane xmlns="http://javafx.com/javafx"
-            xmlns:fx="http://javafx.com/fxml"
-            fx:controller="controllers.QuadTreeController"
-            stylesheets="@/styles/quadtree.css">
-
-    <center>
-
-        <StackPane fx:id="sizePane" minHeight="600" maxHeight="600" minWidth="600" maxWidth="600" prefWidth="600" prefHeight="600">
-
-            <Canvas fx:id="quadTreeCanvas" height="600" width="600" onMouseClicked="#onMouseClicked"
-                    onMouseMoved="#onMouseMoved" onMouseDragged="#onMouseDragged"/>
-
-            <Canvas fx:id="queryCanvas" height="600" width="600" mouseTransparent="true"/>
-
-            <BorderPane fx:id="queryArea" mouseTransparent="true" managed="false"/>
-
-        </StackPane>
-
-    </center>
-
-</BorderPane>
